pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Jayden", email = "jpylesbusiness@gmail.com"},
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.26.0",
    "edgedb>=1.8.0",
    "bcrypt>=4.1.2",
    "pyjwt>=2.8.0",
]
requires-python = ">=3.11"
license = {text = "MIT"}
